@@ -6,15 +6,21 @@
     "start": "node ./bin/www"
   },
   "dependencies": {
-    "body-parser": "~1.8.1",
-    "cookie-parser": "~1.3.3",
-    "debug": "~2.0.0",
-    "express": "~4.9.0",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-register": "^6.9.0",
+    "body-parser": "^1.8.4",
+    "cookie-parser": "^1.3.5",
+    "debug": "^2.0.0",
+    "express": "^4.9.8",
     "hjs": "~0.0.6",
-    "jsx-loader": "^0.12.2",
-    "morgan": "~1.3.0",
-    "node-jsx": "^0.12.4",
-    "react": "^0.12.2",
-    "serve-favicon": "~2.1.3"
+    "morgan": "^1.3.2",
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
+    "serve-favicon": "^2.1.7"
+  },
+  "devDependencies": {
+    "babel-loader": "^6.2.4",
+    "webpack": "^1.13.1"
   }
 }
