@@ -1,8 +1,5 @@
-var React = require("react"),
-    App = React.createFactory(require("components/app"));
-    
-if (typeof window !== "undefined") {
-  window.onload = function() {
-    React.render(App(), document.getElementById("content"));
-  };
-}
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "components/app";
+
+ReactDOM.render(<App />, document.getElementById("content"));
