@@ -1,19 +1,21 @@
-var React = require("react");
+import React from "react";
 
-var Search = React.createClass({
-  getInitialState() {
-    return {
+export default class Search extends React.Component {
+  constructor() {
+    super();
+
+    this.state = {
       search: ""
     };
-  },
+  }
   render() {
     return (
       <div className="search-component">
-        <input type="text" onChange={this.changeSearch} />
+        <input type="text" onChange={this.changeSearch.bind(this)} />
         <p><span>You are searching for: {this.state.search}</span></p>
       </div>
     );
-  },
+  }
   changeSearch(event) {
     var text = event.target.value;
 
@@ -21,6 +23,4 @@ var Search = React.createClass({
       search: text
     });
   }
-});
-
-module.exports = Search;
+}
