@@ -9,7 +9,7 @@ module.exports = [{
   },
   module: {
     loaders: [
-      { test: /\.jsx$/, loader: "jsx-loader?harmony"}
+      { test: /\.jsx?$/, loader: "babel-loader"}
     ]
   },
   resolve: {
